Ändrat på jobbet, men kan inte provköra
Jag har inte fått till projektinställningarna, så att junit hittas.
Det syns inte i Git-repot.

+++ ED1U3/src/helgestenstrom/D1P3C/Bag.java
@@ -13,7 +13,7 @@ public class Bag {
         this.length = length;
     }
 
-    public double volume() {
+    public int volume() {
         return height * width * length;
     }
